feat(tool): 更新工具管理服务，重构用户工具实体，添加工具版本支持及卸载功能

+++ AgentX/src/main/java/org/xhy/interfaces/dto/Page.java
@@ -2,23 +2,23 @@ package org.xhy.interfaces.dto;
 
 public class Page {
 
-    private Long page = 1L;
+    private Integer page = 1;
 
-    private Long pageSize  = 15L;
+    private Integer pageSize  = 15;
 
-    public Long getPage() {
+    public Integer getPage() {
         return page;
     }
 
-    public void setPage(Long page) {
+    public void setPage(Integer page) {
         this.page = page;
     }
 
-    public Long getPageSize() {
+    public Integer getPageSize() {
         return pageSize;
     }
 
-    public void setPageSize(Long pageSize) {
+    public void setPageSize(Integer pageSize) {
         this.pageSize = pageSize;
     }
 }
